+++ app/src/main/java/pl/rethagos/musicplayer/recyclerview/OnFolderPathClickListener.kt
@@ -1,0 +1,7 @@
+package pl.rethagos.musicplayer.recyclerview
+
+import pl.rethagos.musicplayer.model.FolderPath
+
+interface OnFolderPathClickListener {
+    fun onClick(folderPath: FolderPath?)
+}